Isolate bank bundle simulation
